Moved FWB JS to main routes directory
