+++ cfgov/wellbeing/jinja2/wellbeing/home.html
@@ -982,94 +982,3 @@
         </a>
     </div>
 {%- endblock %}
-
-{% block javascript %}
-    {{ super() }}
-
-    <script>
-    /* eslint no-global-assign: "off" */
-    'use strict';
-
-    ( function() {
-
-      var totalQuestions = 12;
-      var quizComplete = false;
-      var questionValues = {};
-      var radioButtons = document.querySelectorAll( '[type="radio"]' );
-      var submitBtn = document.querySelector( '#submit-quiz' );
-
-      function allQuestionsCompleted() {
-        return Object.keys( questionValues ).length === totalQuestions;
-      }
-
-      function enableSubmit() {
-        submitBtn.title = 'Get your score';
-        var className = 'a-btn__disabled';
-
-        if ( submitBtn.classList ) {
-          submitBtn.classList.remove( className );
-        } else {
-          // Support browsers who don't have the classList API.
-          submitBtn.className = submitBtn.className
-                                         .replace( new RegExp( '(^|\\b)' + className.split( ' ' )
-                                                  .join( '|' ) + '(\\b|$)', 'gi' ), ' ' );
-        }
-      }
-
-      function handleRadio( input ) {
-        if ( input.name && input.checked ) {
-          questionValues[input.name] = input.value;
-        }
-
-        quizComplete = allQuestionsCompleted();
-
-        if ( quizComplete ) {
-          enableSubmit();
-        }
-      }
-
-      function sendEvent( action, label, category ) {
-        var eventData = Analytics.getDataLayerOptions( action, label, category );
-        Analytics.sendEvent( eventData );
-      }
-
-      submitBtn.addEventListener( 'click', function( event ) {
-        // Confirm that all questions are answered
-        if ( !quizComplete ) {
-          // If not, block the form submission action
-          return event.preventDefault();
-        }
-        // Otherise, allow the form submission to go through as normal
-        return null;
-      } );
-
-      [].forEach.call( radioButtons, function( el ) {
-        el.addEventListener( 'click', function( event ) {
-          var input = event.target;
-          handleRadio( input );
-
-          var action = input.getAttribute( 'data-gtm-action' );
-          var label = input.getAttribute( 'data-gtm-label' );
-          var category = input.getAttribute( 'data-gtm-category' );
-
-          if ( Analytics.tagManagerIsLoaded ) {
-            sendEvent( action, label, category );
-          } else {
-            Analytics.addEventListener( 'gtmLoaded', sendEvent );
-          }
-        } );
-      } );
-
-      // Disable submit button initially
-      submitBtn.classList.add( 'a-btn__disabled' );
-      submitBtn.title = 'Please answer all questions to get your score';
-
-      // Look at all radios on coming back to the page or on reload
-      radioButtons.forEach( function( input ) {
-        handleRadio( input );
-      } );
-
-    } )();
-
-    </script>
-{% endblock %}
